+++ src/main/scala/de/thatsich/minecraft/intellie/applied/aerodynamics/proxy/module/disassembler/item/config/DisassemblerWeaponConfigAccess.scala
@@ -12,9 +12,9 @@ import de.thatsich.minecraft.common.config.Config
  */
 class DisassemblerWeaponConfigAccess(config: Config) extends DisassemblerWeaponConfig
 {
-	override val minimalDamageVsEntites: Float = this.config.getFloat("Combat", "initialDamageVsEntites", 0)
+	override val minimalDamageVsEntites: Double = this.config.getDouble("Combat", "initialDamageVsEntites", 0)
 
-	override val maximalDamageVsEntites: Float = this.config.getFloat("Combat", "maximalDamageVsEntites", 20)
+	override val maximalDamageVsEntites: Double = this.config.getDouble("Combat", "maximalDamageVsEntites", 20)
 
 	this.config.save()
 }
