+++ src/main/scala/de/thatsich/minecraft/intellie/applied/aerodynamics/proxy/module/suite/freerunner/item/config/FreeRunnerFunctionalityConfigAccess.scala
@@ -13,10 +13,10 @@ import de.thatsich.minecraft.common.config.Config
 private[freerunner] class FreeRunnerFunctionalityConfigAccess(config: Config)
 extends FreeRunnerFunctionalityConfig
 {
-	val minimalWalkSpeed = this.config.getFloat("Functionality", "minimalWalkSpeed", 0.1F)
-	val minimalRunSpeed = this.config.getFloat("Functionality", "minimalRunSpeed", 0.15F)
-	val maximalRunSpeed = this.config.getFloat("Functionality", "maximalRunSpeed", 0.2F)
-	val maximalWalkSpeed = this.config.getFloat("Functionality", "maximalWalkSpeed", 0.11F)
+	val minimalWalkSpeed = this.config.getDouble("Functionality", "minimalWalkSpeed", 0.1)
+	val minimalRunSpeed = this.config.getDouble("Functionality", "minimalRunSpeed", 0.15)
+	val maximalRunSpeed = this.config.getDouble("Functionality", "maximalRunSpeed", 0.2)
+	val maximalWalkSpeed = this.config.getDouble("Functionality", "maximalWalkSpeed", 0.11)
 
 	this.config.save()
 }
